ajustes

+++ app/Http/Controllers/Tenant/VoidedController.php
@@ -127,7 +127,7 @@ class VoidedController extends Controller
         });
         $voided_doc = VoidedDocument::where('id', $voided_id)->first();
         $doc_payment_seat = DocumentPayment::where('document_id', $voided_doc->document_id)->first();
-        if($doc_payment_seat->count() > 0 ){
+        if($doc_payment_seat && $doc_payment_seat->count() > 0 ){
             AccountingEntries::where('document_id', 'CF'.$doc_payment_seat->id)->delete();
         }
